pyproject.toml
[tool.poetry]
name = "m-19-prod"
version = "0.1.0"
description = ""
authors = ["Alexandr Basso <alexandr.basso@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.0"
djangorestframework = "^3.14.0"
asgiref = "^3.7.2"
attrs = "^23.1.0"
certifi = "^2023.5.7"
colorama = "^0.4.6"
django-debug-toolbar = "^4.1.0"
django-filter = "^23.2"
docutils = "^0.20.1"
drf-spectacular = "^0.26.3"
gunicorn = "^20.1.0"
inflection = "^0.5.1"
jsonschema = "^4.18.0"
parts = "^1.6.0"
pillow = "^10.0.0"
products = "^1.2.0"
pyrsistent = "^0.19.3"
pytz = "^2023.3"
pyyaml = "^6.0"
sentry-sdk = "^1.28.0"
sqlparse = "^0.4.4"
typing-extensions = "^4.7.1"
tzdata = "^2023.3"
uritemplate = "^4.1.1"
urllib3 = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
